[TUI] Rearrange key table, fix layout operators
(<++>), (<==>) don't divide the available space proportionally. When
joining multiple components it is vital to combine them in one hBox/vBox
call.

Now there is hBox' and vBox' which insert the spacer element between
elements. (<++>) and (<==>) are defined in terms of these functions.

BUT: the binary operations still give bad results when nested. Use the
list functions instead!